Create a display menu options file
