+++ main.rb
@@ -1,4 +1,5 @@
 require './app'
+require './display_menu_options'
 
 def main
   puts
@@ -20,17 +21,6 @@ def main_extended
 
     puts
   end
-end
-
-def display_menu_options
-  puts 'Please choose an option by entering a number:'
-  puts '1 - List all books'
-  puts '2 - List all people'
-  puts '3 - Create a person'
-  puts '4 - Create a book'
-  puts '5 - Create a rental'
-  puts '6 - List all rentals for a given person id'
-  puts '7 - Exit'
 end
 
 def handle_user_input(user_input, app)
